requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-countries==6.1.2
django-phone-field==1.8.1
gunicorn==20.0.4
idna==2.10
numpy==1.19.1
pandas==1.1.0
pgeocode==0.2.1
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
